apiserver/requirements.txt
django==2.2.8
django-cors-middleware==1.3.1
django-extensions==2.2.5
djangorestframework>=3.9.1
PyJWT==1.4.2
six==1.10.0
Pillow
